Fixed a numerical instability in the virial stepping

+++ src/dynamo/dynamo/interactions/potentials/lennard_jones.hpp
@@ -72,7 +72,7 @@ namespace dynamo {
 
     virtual void outputXML(magnet::xml::XmlStream&) const;
 
-    long double B2func(long double) const;
+    double B2func(double) const;
 
     double _sigma;
     double _epsilon;
